Move some stuff around to document more refactors
