+++ internal/k8s/reconciler/gateway.go
@@ -3,8 +3,6 @@ package reconciler
 import (
 	"context"
 	"encoding/json"
-	"fmt"
-	"strings"
 	"sync/atomic"
 
 	gw "sigs.k8s.io/gateway-api/apis/v1alpha2"
@@ -125,57 +123,4 @@ func (g *K8sGateway) TrackSync(ctx context.Context, sync func() (bool, error)) e
 		}
 	}
 	return nil
-}
-
-func (g *K8sGateway) ensureDeploymentExists(ctx context.Context) error {
-	// Create service account for the gateway
-	if serviceAccount := g.config.ServiceAccountFor(g.Gateway); serviceAccount != nil {
-		if err := g.client.EnsureServiceAccount(ctx, g.Gateway, serviceAccount); err != nil {
-			return err
-		}
-	}
-
-	deployment := g.deploymentBuilder.Build()
-	mutated := deployment.DeepCopy()
-	if updated, err := g.client.CreateOrUpdateDeployment(ctx, mutated, func() error {
-		mutated = apigwv1alpha1.MergeDeployment(deployment, mutated)
-		return g.client.SetControllerOwnership(g.Gateway, mutated)
-	}); err != nil {
-		return fmt.Errorf("failed to create or update gateway deployment: %w", err)
-	} else if updated {
-		if g.logger.IsTrace() {
-			data, err := json.MarshalIndent(mutated, "", "  ")
-			if err == nil {
-				g.logger.Trace("created or updated gateway deployment", "deployment", string(data))
-			}
-		}
-	}
-
-	// Create service for the gateway
-	if service := g.serviceBuilder.Build(); service != nil {
-		mutated := service.DeepCopy()
-		if updated, err := g.client.CreateOrUpdateService(ctx, mutated, func() error {
-			mutated = apigwv1alpha1.MergeService(service, mutated)
-			return g.client.SetControllerOwnership(g.Gateway, mutated)
-		}); err != nil {
-			return fmt.Errorf("failed to create or update gateway service: %w", err)
-		} else if updated {
-			if g.logger.IsTrace() {
-				data, err := json.MarshalIndent(mutated, "", "  ")
-				if err == nil {
-					g.logger.Trace("created or updated gateway service", "service", string(data))
-				}
-			}
-		}
-	}
-
-	return nil
-}
-
-func setToCSV(set map[string]struct{}) string {
-	values := []string{}
-	for value := range set {
-		values = append(values, value)
-	}
-	return strings.Join(values, ", ")
 }
